+++ src/ts/Ads/Views/EndScreen.ts
@@ -4,7 +4,7 @@ import { IGDPREventHandler } from 'Ads/EventHandlers/GDPREventHandler';
 import { AdUnitStyle } from 'Ads/Models/AdUnitStyle';
 import { CustomFeatures } from 'Ads/Utilities/CustomFeatures';
 import { AbstractPrivacy, IPrivacyHandlerView } from 'Ads/Views/AbstractPrivacy';
-import { ABGroup } from 'Core/Models/ABGroup';
+import { ABGroup, AccessibleCloseButtonTest } from 'Core/Models/ABGroup';
 import { Localization } from 'Core/Utilities/Localization';
 import { View } from 'Core/Views/View';
 import EndScreenTemplate from 'html/EndScreen.html';
@@ -117,6 +117,11 @@ export abstract class EndScreen extends View<IEndScreenHandler> implements IPriv
         // Android <= 4.4.4
         if (this._platform === Platform.ANDROID && this._apiLevel! <= 19) {
             this._container.classList.add('old-androids');
+        }
+
+        // Close Button Experiment
+        if (AccessibleCloseButtonTest.isValid(this._abGroup)) {
+            document.documentElement.classList.add('accessible-close-button');
         }
     }