+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Pending
+
+* Experiment to test more accessible mraid close button in playables [#8614](https://github.com/Applifier/unity-ads-webview/pull/8614)
+* Show GDPR banner also in Performance VideoOverlay [#8637](https://github.com/Applifier/unity-ads-webview/pull/8637)
+* Add Reverse AB Test for "install now" button in rewarded ads [#8522](https://github.com/Applifier/unity-ads-webview/pull/8522)
+* Add Accessible close button test [#8471](https://github.com/Applifier/unity-ads-webview/pull/8471)
+
 # Tue Mar 12 21:19:07 UTC 2019
 
 * Revert "Don't send operative clicks for Programmatic MRAID" [#8639](https://github.com/Applifier/unity-ads-webview/pull/8639)